feat(graphics backend): Replaced `API` and all related code with `_Backend`, `GraphicsBackend`, and `WindowingBackend`; also added `BackendManager` for managing the current graphics backend.
I separated the graphics API abstraction into "graphics" and "windowing" pieces mainly because of one reason: There are multiple different window context libraries for every graphics API (e.g: GLFW isn't the only OpenGL window context library). Plus, you could make your own (which I am NOT going to do, but *you* totally could).

Interfacing with the graphics API & window context library is done through `BackendManager` via the `global_BackendManager` singleton accessor. The exact interfacing implementation isn't set in stone, yet, which is why most of the functions are labelled as prototypes. Input handling is what I want to attack next, as the only input handling, currently, is a debug/prototype "exit on escape key pressed" function.

One idea I have is to be able to store the results of the GLFW key callback function in a format that's able to be parsed and interfaced with as efficiently as I can make it.

That, or the windowing backend will just pass inputs to the input manager directly, which is less graceful but more simple/direct.