feature(input manager and glfw backend interfacing with keyboard handler): InputManager initializes the global KeyboardHandler in its own `Init` function. I also wrote a prototype key callback function that implements my custom KeyID & KeyAction types. Currently, it's effectively just an abstraction of the GLFW key callback function, but if I ever want to integrate different windowing libraries, this function will let me do so without having to write/rewrite all the code for handling inputs.
The GLFW backend uses the GLFW to KeyID conversion map to translate GLFW key callbacks into KeyID callbacks and then passes them to the KeyHandler. The KeyHandler is what finally passes the information to the global InputManager. I plan on ironing this code out some more as it's pretty wrinkly and I feel like it's got some redundant parts to it.