feature(commands): Created the `CommandLine`, which also stores the different console commands (which are just strings), as well as a map that pairs command strings with their function pointers.
This is how commands are executed: `CommandLine::try_RunCommand` is passed a command (string) and safely tries to call the function pointer mapped to that string (returning 'Status::CommandLineINVALID_COMMAND' if the passed string is not in the map).

I've also implemented a queue into 'CommandLine', as well as functions for processing the queue. However, I'm considering removing this feature in favor of executing commands directly and immediately; since the only real reason for processing commands in a queue would be for input events, and those are already processed in a queue, this feature is redundant.