feature: Removed fullscreen function from `RenderingBackend` and overhauled how the backends' settings are changed
`BackendManager` now has a function for updating the windowing backend's state. It's just a function that calls a function in the windowing backend, but it's useful for keeping the user from having to include 'backend.hpp' as well as 'backend_manager.hpp'.

`WindowingBackend` now has functions for moving, resizing, and toggling fullscreen. I've implemented these functions in `GLFW_Backend`. There's also an 'UpdateState' function, that will detect changes to the settings and update the window accordingly.