feature(major): Overhauled 'ResourceManager' and implemented mesh & texture loading functions
ResourceManager is now in control of all resources; this includes meshes and textures. It can safely try to create new resources, and manages multiple vectors of different resources alongside a map of resource ID numbers paired with 'Resource' pointers. This map is what the various games/apps can use to get a Resource that's tied to an ID. The vectors help me avoid using 'new' and 'delete' (as well as smart pointers). There are a few resources that are always loaded (error mesh, missing texture, etc). These have constant, static IDs that do not change and must always be paired with their respective resource. Other than those, the rest of the code will not use static, constexpr resource IDs.