feature: Created the TheatreParser and implemented Theatre file parsing
The parser properly iterates over every character in a Theatre file and parses the data it reads. The bulk of the parser is contained within a large switch statement, for efficiency purposes. It's pretty much a giant state machine, where the main state is what the current character is. I intend on cleaning it up and trying to optimize it (especially the beginning with the extra switch statement and extraneous if statements). Hopefully it's self-documenting enough, because I really don't feel like writing a large description about it right now.