feature: Created the 'BinaryFileData' struct
It used to be a more robust struct that stored data as `unsigned char*` (with an `unsigned int` size), but as handling a pointer array like that became more annoying and cumbersome, once I learned you could just dump the pointer array into a vector I immediately opted to use vectors instead. This made my life so much easier.