feature: Implemented a scrappy set of function templates for converting a string to a number
This implementation considers the string to be in the format of a "Numeric" `variable_t` value. It's mainly just a way for me to nicely and safely use `std::stoX` without having to re-write code for every necessary version of `std::stoX` (mainly, `std::stol` for integral types, and `std::stod` for real types).