feature(test_app): Removed the resource manager from the main app and added lots of debug features
The debugger will now log theater load/unload times and display them all on a separate window (the logs can be cleared with a button). It also has a sort of "stopwatch" that can either be manually triggered or triggered in the code via global variables; the stopwatch can also be prevented from being triggered in the code with a button (this state is a toggle). Stopwatch times are logged in a separate window and the logs can be cleared with a button.