feature: Added 'AddID', 'Contains', and 'EndOfReservedIDs' to 'ID'
'AddID' attempts to add the passed ID to 's_ExistingIDs' and returns true if successful or false if unsuccessful.
'Contains' is a wrapper for 's_ExistingIds.contains(id)'.
'EndOfReservedIDs' is a `constexpr uint` that lets me define a range of numbers from 0 to 'EndOfReservedIDs' which the random generator will never pick.