refactor: Mesh data is buffered and obtained more safely.
Mesh data is now obtained via a function that safely returns the data buffered with 'Models::ID::Error' if the passed id isn't in the map. The Error mesh will always be buffered at 'Models::ID::Error'. When Mesh data is buffered, if the Resource's status is not "successful", no mesh data is buffered (leading to the Error Mesh being used).