refactor(major): Moved 'GraphicsBackend' and 'WindowingBackend' into their own header files and refactored 'BackendManager'
Now, 'BackendManager' returns smart pointers to the current backends instead of their IDs, and cannot swap backends on during runtime (which was impossible, anyways). It also includes both the header files for 'GraphicsBackend' and 'WindowingBackend', so if you need both/all three, you only need to include the one header. This is how it worked previously.