feature(breaking): Replaced 'Key/KeyID' and 'KeyBind' with 'InputID' and 'binding_t'
Input IDs for all inputs are handled by 'InputID', using the 'frozen' header library for compile-time containers. "Bindings" (binding_t) are small wrappers for an InputID, InputStatus, and booleans for lock/"just changed" status. 'InputStatus' is a simple enum with two states (it was less confusing to use an enum instead of a boolean).